blueprints/agentic/multi-agent-strands-bedrock/citymapper/pyproject.toml
[project]
name = "citymapper-travel-agent"
version = "0.1.0"
description = "Citymapper Travel Agent with activities and mapping capabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.8",
    "mcp[cli]>=1.9.4",
    "strands-agents[a2a]>=0.1.9",
    "strands-agents-tools>=0.1.6",
    "markdown",
    "rich",
    "uv==0.7.11",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "aiohttp>=3.8.0",
    "requests>=2.31.0",
    "pyjwt==2.10.1",
    "cryptography==45.0.4",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0"
]

[project.scripts]
"mcp-server"     = "main:main_mcp_server"
"a2a-server"     = "main:main_a2a_server"
"fastapi-server" = "main:main_fastapi"
"interactive"    = "main:main_interactive"
"agent"          = "main:servers"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]